pyproject.toml
[project]
name = "agents2-s4-paint-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-api-python-client>=2.184.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "google-genai>=1.41.0",
    "google-generativeai>=0.8.5",
    "mcp[cli]>=1.16.0",
    "pyautogui>=0.9.54",
    "python-dotenv>=1.1.1",
    "python-pptx>=1.0.2",
    "pywinauto>=0.6.9",
]
